+++ http4k-multipart/src/main/kotlin/org/http4k/multipart/MultipartFormBuilder.kt
@@ -1,18 +1,19 @@
 package org.http4k.multipart
 
+import org.http4k.core.Parameters
+import org.http4k.lens.MULTIPART_BOUNDARY
 import java.io.InputStream
 import java.io.SequenceInputStream
 import java.nio.charset.Charset
 import java.nio.charset.StandardCharsets
-import java.util.ArrayDeque
-import java.util.Collections
+import java.util.*
 
 internal class MultipartFormBuilder(inBoundary: ByteArray, private val encoding: Charset = Charset.defaultCharset()) {
     private val boundary = ArrayDeque<ByteArray>()
 
     private val waitingToStream = mutableListOf<InputStream>()
 
-    constructor(boundary: String) : this(boundary.toByteArray(StandardCharsets.UTF_8), StandardCharsets.UTF_8)
+    constructor(boundary: String = MULTIPART_BOUNDARY) : this(boundary.toByteArray(StandardCharsets.UTF_8), StandardCharsets.UTF_8)
 
     init {
         boundary.push(StreamingMultipartFormParts.prependBoundaryWithStreamTerminator(inBoundary))
@@ -26,22 +27,21 @@ internal class MultipartFormBuilder(inBoundary: ByteArray, private val encoding:
         return SequenceInputStream(Collections.enumeration(waitingToStream))
     }
 
-    fun field(name: String, value: String): MultipartFormBuilder = apply {
-        part(value, Pair("Content-Disposition", listOf("form-data" to null, "name" to name)))
+    fun field(name: String, value: String, headers: Parameters): MultipartFormBuilder = apply {
+        part(value, listOf("Content-Disposition" to """form-data; name="$name"""") + headers)
     }
 
-    private fun appendHeader(headerName: String, pairs: List<Pair<String, String?>>) {
-        val headers = "$headerName: " + pairs.joinToString("; ") { (first, second) ->
-            if (second != null) """$first="$second"""" else first
-        }
+    private fun appendHeader(headerName: String, headerValue: String?) {
+        val headerLine = "$headerName: ${headerValue.orEmpty()}"
 
-        add(headers.toByteArray(encoding))
+        add(headerLine.toByteArray(encoding))
         add(StreamingMultipartFormParts.FIELD_SEPARATOR)
     }
 
-    fun part(contents: String, vararg headers: Pair<String, List<Pair<String, String?>>>): MultipartFormBuilder = part(contents.byteInputStream(encoding), *headers)
+    fun part(contents: String, headers: Parameters) =
+        part(contents.byteInputStream(encoding), headers)
 
-    fun part(contents: InputStream, vararg headers: Pair<String, List<Pair<String, String?>>>): MultipartFormBuilder = apply {
+    fun part(contents: InputStream, headers: Parameters) = apply {
         add(boundary.peek())
         add(StreamingMultipartFormParts.FIELD_SEPARATOR)
         if (headers.isNotEmpty()) {
@@ -59,23 +59,27 @@ internal class MultipartFormBuilder(inBoundary: ByteArray, private val encoding:
     fun startMultipart(multipartFieldName: String, subpartBoundary: String): MultipartFormBuilder = apply {
         add(boundary.peek())
         add(StreamingMultipartFormParts.FIELD_SEPARATOR)
-        appendHeader("Content-Disposition", listOf(Pair("form-data", null), Pair("name", multipartFieldName)))
-        appendHeader("Content-Type", listOf(Pair("multipart/mixed", null), Pair("boundary", subpartBoundary)))
+        appendHeader("Content-Disposition", """form-data; name="$multipartFieldName"""")
+        appendHeader("Content-Type", """multipart/mixed; boundary="$subpartBoundary"""")
         add(StreamingMultipartFormParts.FIELD_SEPARATOR)
         boundary.push((String(StreamingMultipartFormParts.STREAM_TERMINATOR, encoding) + subpartBoundary).toByteArray(encoding))
     }
 
-    fun attachment(fileName: String, contentType: String, contents: String): MultipartFormBuilder =
+    fun attachment(fileName: String, contentType: String, contents: String,
+                   headers: Parameters) =
         part(contents,
-            Pair("Content-Disposition", listOf(Pair("attachment", null), Pair("filename", fileName))),
-            Pair("Content-Type", listOf(Pair(contentType, null)))
-        )
+            listOf(
+                "Content-Disposition" to """attachment; filename="$fileName"""",
+                "Content-Type" to contentType) + headers
+            )
 
-    fun file(fieldName: String, filename: String, contentType: String, contents: InputStream): MultipartFormBuilder =
+    fun file(fieldName: String, filename: String, contentType: String, contents: InputStream,
+             headers: Parameters) =
         part(contents,
-            "Content-Disposition" to listOf("form-data" to null, "name" to fieldName, "filename" to filename),
-            "Content-Type" to listOf(contentType to null)
-        )
+             listOf(
+                "Content-Disposition" to """form-data; name="$fieldName"; filename="$filename"""",
+                "Content-Type" to contentType) + headers
+            )
 
     fun endMultipart(): MultipartFormBuilder = apply {
         add(boundary.pop())