+++ http4k-multipart/src/test/kotlin/org/http4k/lens/MultipartFormFieldTest.kt
@@ -7,18 +7,18 @@ import com.natpryce.hamkrest.throws
 import org.junit.jupiter.api.Test
 
 class MultipartFormFieldTest {
-    private val form = MultipartForm(mapOf("hello" to listOf("world", "world2")))
+    private val form = MultipartForm(mapOf("hello" to listOf(MultipartFormField("world"), MultipartFormField("world2"))))
 
     @Test
     fun `value present`() {
-        assertThat(MultipartFormField.optional("hello")(form), equalTo("world"))
-        assertThat(MultipartFormField.required("hello")(form), equalTo("world"))
-        assertThat(MultipartFormField.map { it.length }.required("hello")(form), equalTo(5))
-        assertThat(MultipartFormField.map { it.length }.optional("hello")(form), equalTo(5))
+        assertThat(MultipartFormField.string().optional("hello")(form), equalTo("world"))
+        assertThat(MultipartFormField.string().required("hello")(form), equalTo("world"))
+        assertThat(MultipartFormField.string().map { it.length }.required("hello")(form), equalTo(5))
+        assertThat(MultipartFormField.string().map { it.length }.optional("hello")(form), equalTo(5))
 
         val expected: List<String?> = listOf("world", "world2")
-        assertThat(MultipartFormField.multi.required("hello")(form), equalTo(expected))
-        assertThat(MultipartFormField.multi.optional("hello")(form), equalTo(expected))
+        assertThat(MultipartFormField.string().multi.required("hello")(form), equalTo(expected))
+        assertThat(MultipartFormField.string().multi.optional("hello")(form), equalTo(expected))
     }
 
     @Test
@@ -34,53 +34,53 @@ class MultipartFormFieldTest {
 
     @Test
     fun `value replaced`() {
-        val single = MultipartFormField.required("world")
+        val single = MultipartFormField.string().required("world")
         assertThat(single("value2", single("value1", form)), equalTo(form + ("world" to "value2")))
 
-        val multi = MultipartFormField.multi.required("world")
+        val multi = MultipartFormField.string().multi.required("world")
         assertThat(multi(listOf("value3", "value4"), multi(listOf("value1", "value2"), form)),
             equalTo(form + ("world" to "value3") + ("world" to "value4")))
     }
 
     @Test
     fun `invalid value`() {
-        val requiredFormField = MultipartFormField.map(String::toInt).required("hello")
+        val requiredFormField = MultipartFormField.string().map(String::toInt).required("hello")
         assertThat({ requiredFormField(form) }, throws(lensFailureWith<MultipartForm>(Invalid(requiredFormField.meta), overallType = Failure.Type.Invalid)))
 
-        val optionalFormField = MultipartFormField.map(String::toInt).optional("hello")
+        val optionalFormField = MultipartFormField.string().map(String::toInt).optional("hello")
         assertThat({ optionalFormField(form) }, throws(lensFailureWith<MultipartForm>(Invalid(optionalFormField.meta), overallType = Failure.Type.Invalid)))
 
-        val requiredMultiFormField = MultipartFormField.map(String::toInt).multi.required("hello")
+        val requiredMultiFormField = MultipartFormField.string().map(String::toInt).multi.required("hello")
         assertThat({ requiredMultiFormField(form) }, throws(lensFailureWith<MultipartForm>(Invalid(requiredMultiFormField.meta), overallType = Failure.Type.Invalid)))
 
-        val optionalMultiFormField = MultipartFormField.map(String::toInt).multi.optional("hello")
+        val optionalMultiFormField = MultipartFormField.string().map(String::toInt).multi.optional("hello")
         assertThat({ optionalMultiFormField(form) }, throws(lensFailureWith<MultipartForm>(Invalid(optionalMultiFormField.meta), overallType = Failure.Type.Invalid)))
     }
 
     @Test
     fun `sets value on form`() {
-        val formField = MultipartFormField.required("bob")
+        val formField = MultipartFormField.string().required("bob")
         val withFormField = formField("hello", form)
         assertThat(formField(withFormField), equalTo("hello"))
     }
 
     @Test
     fun `can create a custom type and get and set on request`() {
-        val custom = MultipartFormField.map(::MyCustomBodyType, { it.value }).required("bob")
+        val custom = MultipartFormField.string().map(::MyCustomBodyType, MyCustomBodyType::value).required("bob")
 
         val instance = MyCustomBodyType("hello world!")
         val formWithField = custom(instance, MultipartForm())
 
-        assertThat(formWithField.fields["bob"], equalTo(listOf("hello world!")))
+        assertThat(formWithField.fields["bob"], equalTo(listOf(MultipartFormField("hello world!"))))
 
         assertThat(custom(formWithField), equalTo(MyCustomBodyType("hello world!")))
     }
 
     @Test
     fun `toString is ok`() {
-        assertThat(MultipartFormField.required("hello").toString(), equalTo("Required form 'hello'"))
-        assertThat(MultipartFormField.optional("hello").toString(), equalTo("Optional form 'hello'"))
-        assertThat(MultipartFormField.multi.required("hello").toString(), equalTo("Required form 'hello'"))
-        assertThat(MultipartFormField.multi.optional("hello").toString(), equalTo("Optional form 'hello'"))
+        assertThat(MultipartFormField.string().required("hello").toString(), equalTo("Required form 'hello'"))
+        assertThat(MultipartFormField.string().optional("hello").toString(), equalTo("Optional form 'hello'"))
+        assertThat(MultipartFormField.string().multi.required("hello").toString(), equalTo("Required form 'hello'"))
+        assertThat(MultipartFormField.string().multi.optional("hello").toString(), equalTo("Optional form 'hello'"))
     }
 }