+++ http4k-multipart/src/test/kotlin/org/http4k/multipart/StreamingMultipartFormEncodingTests.kt
@@ -33,11 +33,11 @@ class StreamingMultipartFormEncodingTests {
 
 
     private fun testForm(form: Iterator<StreamingPart>, simpleChar: String, complexChar: String, encoding: Charset) {
-        assertFilePart(form, "file", "foo.tab" + complexChar, "text/whatever" + simpleChar + complexChar, "This is the content of the file" + simpleChar + complexChar, encoding)
-        assertFieldPart(form, "field" + complexChar, "fieldValue" + simpleChar + complexChar, encoding)
-        assertFieldPart(form, "multi", "value1" + simpleChar, encoding)
-        assertFilePart(form, "anotherFile", "BAR.tab", "text/something" + simpleChar, "This is another file" + simpleChar, encoding)
-        assertFieldPart(form, "multi", "value2" + simpleChar, encoding)
+        assertFilePart(form, "file", "foo.tab$complexChar", "text/whatever$simpleChar$complexChar", "This is the content of the file$simpleChar$complexChar", encoding)
+        assertFieldPart(form, "field$complexChar", "fieldValue$simpleChar$complexChar", encoding)
+        assertFieldPart(form, "multi", "value1$simpleChar", encoding)
+        assertFilePart(form, "anotherFile", "BAR.tab", "text/something$simpleChar", "This is another file$simpleChar", encoding)
+        assertFieldPart(form, "multi", "value2$simpleChar", encoding)
 
         assertThereAreNoMoreParts(form)
     }
@@ -48,11 +48,11 @@ class StreamingMultipartFormEncodingTests {
         val boundaryBytes = boundary.toByteArray(encoding)
         return getMultipartFormParts(boundaryBytes,
             MultipartFormBuilder(boundaryBytes, encoding)
-                .file("file", "foo.tab\uD83D\uDCA9", "text/whatever\u00E9\uD83D\uDCA9", "This is the content of the file\u00E9\uD83D\uDCA9".byteInputStream(encoding))
-                .field("field\uD83D\uDCA9", "fieldValue\u00E9\uD83D\uDCA9")
-                .field("multi", "value1\u00E9")
-                .file("anotherFile", "BAR.tab", "text/something\u00E9", "This is another file\u00E9".byteInputStream(encoding))
-                .field("multi", "value2\u00E9")
+                .file("file", "foo.tab\uD83D\uDCA9", "text/whatever\u00E9\uD83D\uDCA9", "This is the content of the file\u00E9\uD83D\uDCA9".byteInputStream(encoding), emptyList())
+                .field("field\uD83D\uDCA9", "fieldValue\u00E9\uD83D\uDCA9", emptyList())
+                .field("multi", "value1\u00E9", emptyList())
+                .file("anotherFile", "BAR.tab", "text/something\u00E9", "This is another file\u00E9".byteInputStream(encoding), emptyList())
+                .field("multi", "value2\u00E9", emptyList())
                 .stream(), encoding)
     }
 }