+++ http4k-multipart/src/test/kotlin/org/http4k/multipart/StreamingMultipartFormSadTests.kt
@@ -8,7 +8,7 @@ import com.natpryce.hamkrest.present
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.fail
-import java.util.NoSuchElementException
+import java.util.*
 
 class StreamingMultipartFormSadTests {
 
@@ -34,8 +34,8 @@ class StreamingMultipartFormSadTests {
     fun failsWhenGettingNextPastEndOfParts() {
         val boundary = "-----1234"
         val form = getMultipartFormParts(boundary, MultipartFormBuilder(boundary)
-            .file("aFile", "file.name", "application/octet-stream", "File contents here".byteInputStream())
-            .file("anotherFile", "your.name", "application/octet-stream", "Different file contents here".byteInputStream()).stream())
+            .file("aFile", "file.name", "application/octet-stream", "File contents here".byteInputStream(), emptyList())
+            .file("anotherFile", "your.name", "application/octet-stream", "Different file contents here".byteInputStream(), emptyList()).stream())
 
         form.next() // aFile
         form.next() // anotherFile
@@ -52,8 +52,8 @@ class StreamingMultipartFormSadTests {
     fun failsWhenGettingNextPastEndOfPartsAfterHasNext() {
         val boundary = "-----1234"
         val form = getMultipartFormParts(boundary, MultipartFormBuilder(boundary)
-            .file("aFile", "file.name", "application/octet-stream", "File contents here".byteInputStream())
-            .file("anotherFile", "your.name", "application/octet-stream", "Different file contents here".byteInputStream()).stream())
+            .file("aFile", "file.name", "application/octet-stream", "File contents here".byteInputStream(), emptyList())
+            .file("anotherFile", "your.name", "application/octet-stream", "Different file contents here".byteInputStream(), emptyList()).stream())
 
         form.next() // aFile
         form.next() // anotherFile
@@ -72,9 +72,9 @@ class StreamingMultipartFormSadTests {
     fun partHasNoHeaders() {
         val boundary = "-----2345"
         val form = getMultipartFormParts(boundary, MultipartFormBuilder(boundary)
-            .field("multi", "value0")
-            .part("" + CR_LF + "value with no headers")
-            .field("multi", "value2")
+            .field("multi", "value0", emptyList())
+            .part("" + CR_LF + "value with no headers", emptyList())
+            .field("multi", "value2", emptyList())
             .stream())
 
         form.next()
@@ -92,8 +92,8 @@ class StreamingMultipartFormSadTests {
         val boundary = "-----2345"
         val form = getMultipartFormParts(boundary, MultipartFormBuilder(boundary)
             .part("contents of StreamingPart",
-                "Content-Disposition" to listOf("form-data" to null, "bit" to "first", "name" to "first-name"),
-                "Content-Disposition" to listOf("form-data" to null, "bot" to "second", "name" to "second-name"))
+                listOf("Content-Disposition" to "form-data; bit=\"first\"; name=\"first-name\"",
+                    "Content-Disposition" to "form-data; bot=\"second\"; name=\"second-name\""))
             .stream())
 
         val StreamingPart = form.next()
@@ -202,7 +202,7 @@ class StreamingMultipartFormSadTests {
         val chars = CharArray(StreamingMultipartFormParts.HEADER_SIZE_MAX)
         chars.fill('x')
         val form = getMultipartFormParts(boundary, MultipartFormBuilder(boundary)
-            .file("aFile", String(chars), "application/octet-stream", "File contents here".byteInputStream()).stream())
+            .file("aFile", String(chars), "application/octet-stream", "File contents here".byteInputStream(), emptyList()).stream())
 
         assertParseErrorWrapsTokenNotFound(form, "Didn't find end of Token <<\r\n>> within 10240 bytes")
     }
@@ -215,18 +215,18 @@ class StreamingMultipartFormSadTests {
         chars.fill('x')
         val form = getMultipartFormParts(boundary, MultipartFormBuilder(boundary)
             .part("some contents",
-                "Content-Disposition" to listOf("form-data" to null, "name" to "fieldName", "filename" to "filename"),
-                "Content-Type" to listOf("text/plain" to null),
-                "extra-1" to listOf(String(chars) to null),
-                "extra-2" to listOf(String(chars) to null),
-                "extra-3" to listOf(String(chars) to null),
-                "extra-4" to listOf(String(chars) to null),
-                "extra-5" to listOf(String(chars) to null),
-                "extra-6" to listOf(String(chars) to null),
-                "extra-7" to listOf(String(chars) to null),
-                "extra-8" to listOf(String(chars) to null),
-                "extra-9" to listOf(String(chars) to null),
-                "extra-10" to listOf(String(chars, 0, 816) to null) // header section exactly 10240 bytes big!
+                listOf("Content-Disposition" to "form-data; name=\"fieldName\"; filename=\"filename\"",
+                    "Content-Type" to "text/plain",
+                    "extra-1" to String(chars),
+                    "extra-2" to String(chars),
+                    "extra-3" to String(chars),
+                    "extra-4" to String(chars),
+                    "extra-5" to String(chars),
+                    "extra-6" to String(chars),
+                    "extra-7" to String(chars),
+                    "extra-8" to String(chars),
+                    "extra-9" to String(chars),
+                    "extra-10" to String(chars, 0, 816)) // header section exactly 10240 bytes big!
             ).stream())
 
         assertParseErrorWrapsTokenNotFound(form, "Didn't find end of Header section within 10240 bytes")