+++ src/docs/cookbook/multipart_forms/example_processing.kt
@@ -3,7 +3,6 @@ package cookbook.multipart_forms
 import org.http4k.client.ApacheClient
 import org.http4k.core.Body
 import org.http4k.core.ContentType
-import org.http4k.core.FormFile
 import org.http4k.core.Method
 import org.http4k.core.MultipartEntity
 import org.http4k.core.Request
@@ -51,7 +50,7 @@ fun main() {
 
 private fun buildValidMultipartRequest(): Request {
     // define fields using the standard lens syntax
-    val nameField = MultipartFormField.map(::AName, AName::value).required("name")
+    val nameField = MultipartFormField.string().map(::AName, AName::value).required("name")
     val imageFile = MultipartFormFile.optional("image")
 
     // add fields to a form definition, along with a validator
@@ -59,6 +58,6 @@ private fun buildValidMultipartRequest(): Request {
 
     val multipartform = MultipartForm().with(
         nameField of AName("rita"),
-        imageFile of FormFile("image.txt", ContentType.OCTET_STREAM, "somebinarycontent".byteInputStream()))
+        imageFile of MultipartFormFile("image.txt", ContentType.OCTET_STREAM, "somebinarycontent".byteInputStream()))
     return Request(Method.POST, "http://localhost:8000").with(strictFormBody of multipartform)
 }